requirements.dev.txt
numpy~=1.22.2
pandas~=1.4.0
matplotlib~=3.5.1
seaborn~=0.12.2
biopython~=1.79
ete3~=3.1.2
tqdm~=4.62.3
PyYAML~=6.0
pytest~=7.1.2
click~=8.1.3
# scipy~=1.8.1
# statsmodels~=0.13.2
pytest
pytest-cov